attachment test should be at SubmissionForm level
